This is my first task assigned by  CipherByte
The task is about to create a Digital Calculator using C++ programming. Where my program can do the calculation just like a normal calci. But it is not so versatile to perform significant calculations such as sin,cos,tan. But further i will try to add more operations such as square root of a number ,power calculations, and factorial.